requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
DateTime==5.5
greenlet==3.0.3
idna==3.6
importlib-metadata==6.7.0
numpy==1.21.6
pandas==1.3.5
python-dateutil==2.9.0.post0
python.dotenv==0.21.1
pytz==2024.1
requests==2.31.0
six==1.16.0
SQLAlchemy==1.4.46
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
zope.interface==6.2